Agent_Frameworks/LlamaIndex/requirements.txt
# ── LlamaIndex core ────────────────────────────────────────────────────────────
llama-index-core>=0.10.0

# ── LLM & embedding integrations ──────────────────────────────────────────────
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0

# ── Agent (OpenAI function-calling) ───────────────────────────────────────────
llama-index-agent-openai>=0.2.0

# ── Search & web tools ────────────────────────────────────────────────────────
ddgs>=0.1.0
wikipedia>=1.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# ── OpenAI base ───────────────────────────────────────────────────────────────
openai>=1.13.0

# ── Configuration & environment ───────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── Terminal display ──────────────────────────────────────────────────────────
rich>=13.0.0

# ── Async HTTP (used by tools) ────────────────────────────────────────────────
aiohttp>=3.9.0
